pyproject.toml
[project]
name = "apontamentos-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.3.4",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.27.0",
  "psycopg[binary]>=3.2.1",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
  "loguru>=0.7.2",
  "pydantic-settings>=2.3.4",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0", "anyio>=4.4.0", "httpx>=0.27.0", "ruff>=0.5.4", "black>=24.4.2", "mypy>=1.10.0"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*", "scripts*"]
